Add files via upload
Changed the function fit_spectrum so that the chisq attribute of the RegionResult objects it produces lists the chi squared for abundencies when the best shift is used (the one with lowest chi squared among all the abundencies). Also changed how fit_spectrum2 works a little. It seems to work better now, which is surprising since I merely restructured the code. There might be more changes as well.