Add files via upload
Have fixed some issues in synther.py, renamed fit_spectrum2 to fit_spectrum_seeking so it has a more descriptive name and also introduced a function that finds the abundance using equivalent widths rather then chi squared. Note that this function may not calculate things properly yet. Also note that while it does give different results, this is not surprising since the method is different. The question is more a matter of how different is reasonable.

Have also changed the function a little to reduce the number of arguments passed around.