Add files via upload
Not sure how to handle equivalent widths, so I made it calculate a few per line (by reducing the interval a little in different ways) and use the mean as the best value and the standard deviation as an error of some kind, but still not sure if that's a good way. The result becomes a bit larger then the one obtained by fitting with chi squared.