Add files via upload
Added a lot of documentation, and a new argument to the fitting functions that allows the caller to set the model file directly. Before it was a "constant" (as in variable that shouldn't be changed) called _MODEL_FILE in synther.py that determined this. This "constant" has been renamed DEFAULT_MODEL_FILE and is used if no model file was passed to the functions.